fix: moved dns resources to the network
