+++ plugins/azure/test/files/network/dnszones.json
