+++ plugins/azure/test/files/network/recordsets.json
